third_party/WebKit/LayoutTests/inspector/console/console-log-object-with-getter-expected.txt

CONSOLE MESSAGE: line 10: [object Object]
Tests that console logging dumps object values defined by getters and allows to expand it.

console-log-object-with-getter.html:10 Object {}foo: Objecta: 1b: 2__proto__: Objectget foo: ()__proto__: Object
